For finding best para
I have iterated over all para, this will take a lot more time. But in reality everytime  we will add new text in database, we will assign new score to it based on top m(for us =10) mistakes. Then the scores of all the texts will be updated periodically based on the mistake list ( or when user is not racing).